Move schema yamls to src/schemas/ directory
