+++ packages/format/src/describe.ts
@@ -1,6 +1,6 @@
 import * as YAML from "yaml";
 
-import { schemaYamls } from "../yamls";
+import { schemaYamls } from "./schemas/yamls";
 
 import type { JSONSchema as JSONSchemaTyped } from "json-schema-typed/draft-2020-12"
